[FIX] web: allow date field to re-apply same value
This commit allows date (i.e. date, datetime & daterange) fields to
apply a value from the props (e.g. coming from an `onchange`), even if
that value is the same as the initial one.

Before this commit, it was not possible due to the fact that the date
service responsible for the reactivity of the field was updating the
input in an incorrect order, causing the field to display the 'input'
value, and not the one enforced by the props.

Task 4978896

closes odoo/odoo#225841

Signed-off-by: Julien Mougenot (jum) <jum@odoo.com>