[FIX] product: duplicated packaging on one-variant product
**PROBLEM**
When creating a new product, with attributes
but only one variant the packaging is duplicated on save.

**STEP TO REPRODUCE**
1. Install inventory, go to the settings
and enable the product packagings option.
2. Create a new product, with one attribute with one possible value,
and with a packaging.
3. save the product and notice the packaging was duplicated.

**CAUSE**
`packaging_ids` is stored on the product.product model (product variant).
`packaging_ids` in the product.template model is a compute field with an inverse.
When creating the product, the inverse function is called, creating a first packaging.
Later in the product template creation, we copy certains fields to to product variant (packaging_ids being one of them).
Those are passed to the template `write()` method, which create another create Command for packaging_ids on the related variant models.

**FIX**
When iterating over the field name, we verify if it's not already present in the template.
If so, we avoid rewritting it to the template.

opw-5050592

closes odoo/odoo#225863

Signed-off-by: Léo Gizard (legi) <legi@odoo.com>